Add framebuffers
- Framebuffer class added, with OpenGL and Vulkan implementations
- GraphicsContext now has derived classes: "FramebufferGC" for rendering to a framebuffer, and "WindowGC" (which adds a swapchain and renders to framebuffer from the swapchain)
- Fix the flipping UV problem (OpenGL seems to flip Y coordinate when rendering to a framebuffer as opposed to the default backbuffer.  Vulkan does not do this (it renders according to viewport in both cases).
This causes issues when using the result of a framebuffer as a texture for later rendering operations, and I do not want the client code to have to do things like "if OpenGL then y-axis is up else if Vulkan then it' down".
The problem is solved by making the Vulkan rendercore behave the same way as the OpenGL one, and it will now render to framebuffer "upside down".  See code comments in VulkanPipeline for details.

other:
- precompile more headers