The beginnings of scene serialization
WIP:  de-serialization not done yet.
To test: Start pikzelated, then File->New, then File->Save As...

Random thoughts:
We are saving things in an "entity major" way, designed for human readability rather than efficiency.  By "entity major" I mean, we write out each entity, and then what components it has.  As opposed to "component major" which is writing out each component and then which entities have that component.
For efficiency (and when we do binary serialization we will want this) it should be "component major" because that's how an ECS is organised.  We want to just dump out (and read back in) the underlying data pools for each component.